Feat: 유저 프로필이미지 컴포넌트를 공용 컴포넌트로 변경
