+++ src/components/UserProfileImage/UserProfileImage.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const profileImage = style({
+  borderRadius: '50%',
+});
